Add landing-component listing/unassign endpoints and guarded componente deletion

+++ Application/src/Services/Componente/DeleteComponenteService.js
@@ -2,10 +2,19 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.DeleteComponenteService = void 0;
 class DeleteComponenteService {
-    constructor(componenteRepository) {
+    constructor(componenteRepository, landingComponenteRepository) {
         this.componenteRepository = componenteRepository;
+        this.landingComponenteRepository = landingComponenteRepository;
     }
     async execute(id_componente) {
+        const componente = await this.componenteRepository.getById(id_componente);
+        if (!componente) {
+            throw new Error("COMPONENTE_NOT_FOUND");
+        }
+        const isAssigned = await this.landingComponenteRepository.existsByComponente(id_componente);
+        if (isAssigned) {
+            throw new Error("COMPONENTE_ASSIGNED");
+        }
         await this.componenteRepository.delete(id_componente);
     }
 }
